% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Digits.R
\docType{data}
\name{Digits}
\alias{Digits}
\title{load 'Digits' data of multiple features}
\format{
a list\describe{
\item{fou}{76 Fourier coefficients of the character shapes.}
\item{fac}{216 profile correlations.}
\item{kar}{64 Karhunen-Loève coefficients.}
\item{pix}{240 pixel averages in 2 x 3 windows.}
\item{zer}{47 Zernike moments.}
\item{mor}{6 morphological features.}
\item{label}{length-\eqn{2000} vector of class labels.}
}
}
\usage{
data(Digits)
}
\description{
UCI >> Multiple Features Data Set
}
\keyword{datasets}
