% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinesubspace_fixed.R
\name{diglocal_fixed}
\alias{diglocal_fixed}
\title{extract local information : fixed dimension}
\usage{
diglocal_fixed(data, k = 10, ndim = 2, center.by.data = TRUE)
}
\description{
use local PCA
}
\examples{
\dontrun{
X    = as.matrix(iris[,1:4])
digs = diglocal_fixed(X)
}

}
