% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinesubspace_variable.R
\name{diglocal_variable}
\alias{diglocal_variable}
\title{extract local information : variable dimension}
\usage{
diglocal_variable(data, k = 10, pcaratio = 0.9, center.by.data = TRUE)
}
\description{
use local PCA
}
\examples{
\dontrun{
X    = as.matrix(iris[,1:4])
digs = diglocal_fixed(X)
}

}
