% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinesubspace_dist_conway.R
\name{dist_conway}
\alias{dist_conway}
\title{Distance using Conway Embedding}
\usage{
dist_conway(list_center, list_stiefel)
}
\description{
Distance using Conway Embedding
}
\examples{
\dontrun{
# example with Iris data
X    = as.matrix(iris[,1:4])
lab  = as.factor(iris[,5])
digs = diglocal_fixed(X, center.by.data=FALSE)

# compute the distance
dobj = dist_conway(digs$center, digs$stiefel)

# embed by MDS
embed2_data   = cmdscale(stats::dist(X), k=2)
embed2_conway = cmdscale(dobj, k=2)

# visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
plot(embed2_data,   col=lab, pch=19, main="Data MDS")
plot(embed2_conway, col=lab, pch=19, main="Graff MDS")
par(opar)
}

}
