% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Embryoid.R
\docType{data}
\name{Embryoid}
\alias{Embryoid}
\title{load 'Embryoid' data}
\format{
a list\describe{
\item{data}{data matrix that is preprocessed with PCA.}
\item{label}{label info}
}
}
\usage{
data(Embryoid)
}
\description{
Embryoid data. Krishnaswamy Lab.
}
\keyword{datasets}
