\name{exp3riment-package}
\alias{exp3riment-package}
\alias{exp3riment}
\docType{package}
\title{\packageTitle{exp3riment}}
\description{\packageDescription{exp3riment}}
\section{Package Content}{\packageIndices{exp3riment}}
\author{\packageAuthor{exp3riment}}
\section{Maintainer}{\packageMaintainer{exp3riment}}
\keyword{package}
