% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_magic.R
\name{lab_magic}
\alias{lab_magic}
\title{My Version of MAGIC}
\usage{
lab_magic(cellxgene, knn = 5, t = 3, apply_sqrt = FALSE)
}
\description{
(n x p) data : cellxgene count matrix : let's not take sparse ones yet.
}
