% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiview_alternating.R
\name{mv_alternating}
\alias{mv_alternating}
\title{Multi-View : Alternating Diffusion}
\usage{
mv_alternating(markov1, markov2, ndim = 2)
}
\description{
for reusability, use markov transition kernels only.
}
