% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiview_integrated.R
\name{mv_integrated}
\alias{mv_integrated}
\title{Multi-View : Integrated Diffusion}
\usage{
mv_integrated(markov1, markov2, ndim = 2)
}
\description{
for reusability, use markov transition kernels only.
}
