% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiview_spherical.R
\name{mv_spherical}
\alias{mv_spherical}
\title{Multi-View : Spherically Aggregated Diffusion}
\usage{
mv_spherical(markov1, markov2, ndim = 2, weight = 0.5)
}
\description{
for reusability, use markov transition kernels only.
}
