% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate_graph.R
\name{phate_graph}
\alias{phate_graph}
\title{Graphical PHATE for Asymmetric Graph}
\usage{
phate_graph(
  input,
  ndim = 2,
  nbdk = 5,
  alpha = 2,
  alg = c("cmds", "mmds"),
  potential = c("log", "sqrt", "none"),
  n_landmark = 100
)
}
\description{
Graphical PHATE for Asymmetric Graph
}
\examples{
\dontrun{
# load the data
data("dolphins", package="exp3riment")
dat_graph = dolphins$igraph
dat_label = dolphins$label

# try PHATE
dol_log_cmds = phate_graph(dat_graph, potential="log", alg="cmds")
dol_log_mmds = phate_graph(dat_graph, potential="log", alg="mmds")

# visualize
require("igraph")
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3))
plot(dat_graph, vertex.color=dat_label, vertex.label=NA,
     main="igraph")
plot(dat_graph, vertex.color=dat_label, vertex.label=NA,
     main="PHATE+CMDS", layout=dol_log_cmds$embedding)
plot(dat_graph, vertex.color=dat_label, vertex.label=NA,
     main="PHATE+MMDS", layout=dol_log_mmds$embedding)
par(opar)
}

}
