% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate_mview_ver1.R
\name{phate_mview_ver1}
\alias{phate_mview_ver1}
\title{MultiView PHATE Version 1 : Distance Merging (2-View Only)}
\usage{
phate_mview_ver1(
  dist1,
  dist2,
  weight = 0.5,
  nbdk = 5,
  ndim = 2,
  alpha = 2,
  alg = c("cmds", "mmds"),
  potential = c("log", "sqrt", "none"),
  n_landmark = 100
)
}
\value{
a list containing\describe{
\item{transition}{transition matrix of optimal kernel.}
\item{embedding}{embedding matrix.}
}
}
\description{
weight=0 means using dist1 only.
}
\examples{
\dontrun{
data(iris)
A = as.matrix(iris[,1:4]) + 0.1*matrix(rnorm(150*4), ncol=4)
B = as.matrix(iris[,1:4]) + 0.1*matrix(rnorm(150*4), ncol=4)

D1 = stats::dist(A)
D2 = stats::dist(B)
output = phate_mview_ver1(D1, D2, weight=0.5)
}




}
