% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate_original.R
\name{phate_original}
\alias{phate_original}
\title{Original Version of PHATE}
\usage{
phate_original(
  data,
  ndim = 2,
  nbdk = 5,
  alpha = 2,
  alg = c("cmds", "mmds"),
  potential = c("log", "sqrt", "none"),
  n_landmark = 100,
  use_landmark = TRUE
)
}
\arguments{
\item{data}{matrix N observations}

\item{ndim}{target dimension}

\item{nbdk}{neighborhood size}

\item{alpha}{decaying kernel parameter}
}
\value{
a list containing\describe{
\item{transition}{transition matrix of optimal kernel.}
\item{embedding}{embedding matrix.}
}
}
\description{
Original Version of PHATE
}
\examples{
\dontrun{
# load the data
data(Embryoid, package="exp3riment")

nsample = 2000
sub_id  = sample(1:nrow(Embryoid$data), nsample)

sub_data  = Embryoid$data[sub_id,]
sub_label = Embryoid$label[sub_id]

distEB    = stats::as.dist(fast_dist(sub_data))
landmark1 = phate_original(distEB, n_landmark=100)$embedding
landmark2 = phate_original(distEB, n_landmark=250)$embedding
landmark3 = phate_original(distEB, n_landmark=500)$embedding

opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3))
plot(landmark1, pch=19, col=sub_label, main="Case 1")
plot(landmark2, pch=19, col=sub_label, main="Case 2")
plot(landmark3, pch=19, col=sub_label, main="Case 3")
par(opar)
}


}
