% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine_transition.R
\name{routine_transition}
\alias{routine_transition}
\title{Construct PHATE-style transition kernel P=D^{-1}A}
\usage{
routine_transition(data, nbdk = 5, alpha = 2)
}
\description{
take (n x p) matrix or 'dist' object
}
