% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USELESS_tphate_aggregation.R
\name{tphate_aggregation}
\alias{tphate_aggregation}
\title{T-PHATE : Distance Aggregation}
\usage{
tphate_aggregation(
  data,
  ndim = 2,
  nbdk = 5,
  alpha = 2,
  temporal = 5,
  weight = 0.5
)
}
\description{
always take an (TxP) input
}
