% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USELESS_tphate_alternating.R
\name{tphate_alternating}
\alias{tphate_alternating}
\title{T-PHATE : Alternating Diffusion}
\usage{
tphate_alternating(data, ndim = 2, nbdk = 5, alpha = 2, temporal = 5)
}
\description{
always take an (TxP) input
}
