% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tphate_hadamard.R
\name{tphate_hadamard}
\alias{tphate_hadamard}
\title{T-PHATE : Integrated Diffusion}
\usage{
tphate_hadamard(data, ndim = 2, nbdk = 5, alpha = 2, temporal = 5)
}
\description{
always take an (TxP) input
}
