% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tphate_hadamardclust.R
\name{tphate_hadamardclust}
\alias{tphate_hadamardclust}
\title{T-PHATE : Hadamard Product + Clustering}
\usage{
tphate_hadamardclust(
  data,
  ndim = 2,
  nbdk = 5,
  alpha = 2,
  temporal = 5,
  decay = 0
)
}
\description{
always take an (TxP) input
}
