% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tphate_integrated.R
\name{tphate_integrated}
\alias{tphate_integrated}
\title{T-PHATE : Integrated Diffusion}
\usage{
tphate_integrated(data, ndim = 2, nbdk = 5, alpha = 2, temporal = 5)
}
\description{
always take an (TxP) input
}
