% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tphate_strategy3.R
\name{tphate_strategy3}
\alias{tphate_strategy3}
\title{Transition PHATE : Strategy 3}
\usage{
tphate_strategy3(data, ndim = 2, nbdk = 5, alpha = 2, symmetrize = TRUE)
}
\description{
finite machine transition
}
