% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tphate_strategy4.R
\name{tphate_strategy4}
\alias{tphate_strategy4}
\title{Transition PHATE : Strategy 4}
\usage{
tphate_strategy4(data, ndim = 2, nbdk = 5, alpha = 2, symmetrize = TRUE)
}
\description{
scale down the voxel activation affinity by transition probability
}
