% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tphate_strategy5.R
\name{tphate_strategy5}
\alias{tphate_strategy5}
\title{Transition PHATE : Strategy 5}
\usage{
tphate_strategy5(data, ndim = 2, nbdk = 5, alpha = 2, scaler = 0.5)
}
\description{
set 1 for same-cluster affinities
otherwise, use log-like decay
}
