% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tphate_strategy7.R
\name{tphate_strategy7}
\alias{tphate_strategy7}
\title{Transition PHATE : Strategy 7}
\usage{
tphate_strategy7(data, ndim = 2, nbdk = 5, alpha = 2, scaler = 0.5)
}
\description{
use log-like temporal decay without clustering information
}
