% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tphate_strategy8.R
\name{tphate_strategy8}
\alias{tphate_strategy8}
\title{Transition PHATE : Strategy 8}
\usage{
tphate_strategy8(data, ndim = 2, nbdk = 5, alpha = 2, nclust = 10)
}
\description{
alternating diffusion : data + cluster-based transition
}
