% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ver2_phate_graphA.R
\name{ver2_phate_graphA}
\alias{ver2_phate_graphA}
\title{Graphical PHATE for using Adjacency}
\usage{
ver2_phate_graphA(
  input,
  ndim = 2,
  nbdk = 5,
  alpha = 2,
  time_step = NULL,
  add_diagonal = TRUE
)
}
\description{
Remove Landmarking Strategy at this point.
}
