% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ver2_phate_graphE.R
\name{ver2_phate_graphE}
\alias{ver2_phate_graphE}
\title{Graphical PHATE for Effective Resistance}
\usage{
ver2_phate_graphE(
  input,
  ndim = 2,
  nbdk = 5,
  alpha = 2,
  time_step = NULL,
  add_diagonal = TRUE
)
}
\description{
Graphical PHATE for Effective Resistance
}
