% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ver2_phate_original.R
\name{ver2_phate_original}
\alias{ver2_phate_original}
\title{Original PHATE Simplified}
\usage{
ver2_phate_original(data, ndim = 2, nbdk = 5, alpha = 2, time_step = NULL)
}
\description{
Original PHATE Simplified
}
