% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ver2_phate_repair.R
\name{ver2_phate_repair}
\alias{ver2_phate_repair}
\title{Original PHATE Simplified + Metric Repair}
\usage{
ver2_phate_repair(data, ndim = 2, nbdk = 5, alpha = 2, time_step = NULL)
}
\description{
Original PHATE Simplified + Metric Repair
}
\examples{
\dontrun{
# load the data
data(Embryoid, package="exp3riment")

nsample = 2000
sub_id  = sample(1:nrow(Embryoid$data), nsample)

sub_data  = Embryoid$data[sub_id,]
sub_label = Embryoid$label[sub_id]

landmark1 = ver2_phate_repair(sub_data)$embedding
landmark2 = ver2_phate_repair(sub_data)$embedding
landmark3 = ver2_phate_repair(sub_data)$embedding

opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3))
plot(landmark1, pch=19, col=sub_label, main="Case 1")
plot(landmark2, pch=19, col=sub_label, main="Case 2")
plot(landmark3, pch=19, col=sub_label, main="Case 3")
par(opar)
}

}
